requirements.txt
aioredis==1.3.1
asgiref==3.3.1
astroid==2.4.2
async-timeout==3.0.1
attrs==20.3.0
autobahn==20.7.1
Automat==20.2.0
cffi==1.14.3
constantly==15.1.0
cryptography==3.2.1
Django==3.1.3
django-cors-headers==3.6.0
djangorestframework==3.12.2
hiredis==1.1.0
hyperlink==20.0.1
idna==2.10
incremental==17.5.0
isort==5.6.4
lazy-object-proxy==1.4.3
mccabe==0.6.1
msgpack==1.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
pylint==2.6.0
PyMySQL==0.10.1
pyOpenSSL==19.1.0
pytz==2020.4
service-identity==18.1.0
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
Twisted==20.3.0
txaio==20.4.1
wrapt==1.12.1
zope.interface==5.2.0
